test(e2e): move upgrade stale setup branches to helpers
